Load Data Fix (Int to Float). Enemies MarkToDie Mode. Enemy Projectile smaller.

+++ Assets/Lucas/Scripts/Stats/BasicEnemy/LoadEnemyData.cs
@@ -40,7 +40,7 @@ public class LoadEnemyData : ReadGoogleSheets
                 _rangedRobotData._damagePerHit = (int)(float.Parse(tempEnemyData[2]) * _basicEnemyData._damagePerHit);
                 _rangedRobotData._attackSpeed = float.Parse(tempEnemyData[3]);
                 _rangedRobotData._givenXP = int.Parse(tempEnemyData[4]);
-                _rangedRobotData._moveSpeed = int.Parse(tempEnemyData[5]);
+                _rangedRobotData._moveSpeed = float.Parse(tempEnemyData[5]);
                 _rangedRobotData._armor = int.Parse(tempEnemyData[6]);
                 _rangedRobotData._elementalResistance = int.Parse(tempEnemyData[7]);
                 _rangedRobotData._technicalResistance = int.Parse(tempEnemyData[8]);
@@ -51,7 +51,7 @@ public class LoadEnemyData : ReadGoogleSheets
                 _pasuKanData._damagePerHit = (int)(float.Parse(tempEnemyData[2]) * _basicEnemyData._damagePerHit);
                 _pasuKanData._attackSpeed = float.Parse(tempEnemyData[3]);
                 _pasuKanData._givenXP = int.Parse(tempEnemyData[4]);
-                _pasuKanData._moveSpeed = int.Parse(tempEnemyData[5]);
+                _pasuKanData._moveSpeed = float.Parse(tempEnemyData[5]);
                 _pasuKanData._armor = int.Parse(tempEnemyData[6]);
                 _pasuKanData._elementalResistance = int.Parse(tempEnemyData[7]);
                 _pasuKanData._technicalResistance = int.Parse(tempEnemyData[8]);
@@ -62,7 +62,7 @@ public class LoadEnemyData : ReadGoogleSheets
                 _gigaRangedRobotData._damagePerHit = (int)(float.Parse(tempEnemyData[2]) * _basicEnemyData._damagePerHit);
                 _gigaRangedRobotData._attackSpeed = float.Parse(tempEnemyData[3]);
                 _gigaRangedRobotData._givenXP = int.Parse(tempEnemyData[4]);
-                _gigaRangedRobotData._moveSpeed = int.Parse(tempEnemyData[5]);
+                _gigaRangedRobotData._moveSpeed = float.Parse(tempEnemyData[5]);
                 _gigaRangedRobotData._armor = int.Parse(tempEnemyData[6]);
                 _gigaRangedRobotData._elementalResistance = int.Parse(tempEnemyData[7]);
                 _gigaRangedRobotData._technicalResistance = int.Parse(tempEnemyData[8]);
